Isolate test env process to prevent user context overwrite
