+++ test/multi_command_flows/workflow_private_registry.go
@@ -76,59 +76,6 @@ func CreateTestBearerCredentialsHome(t *testing.T) string {
 	return homeDir
 }
 
-// realGoCacheEnv returns GOPATH and GOMODCACHE locations outside t.TempDir()-backed HOME dirs.
-// Overriding HOME makes Go default GOPATH to $HOME/go; module files are read-only and break TempDir cleanup.
-func realGoCacheEnv(t *testing.T) (gopath, gomodcache string) {
-	t.Helper()
-
-	realHome, err := os.UserHomeDir()
-	require.NoError(t, err, "failed to get real home dir")
-
-	gopath = os.Getenv("GOPATH")
-	if gopath == "" {
-		gopath = filepath.Join(realHome, "go")
-	}
-
-	gomodcache = os.Getenv("GOMODCACHE")
-	if gomodcache == "" {
-		gomodcache = filepath.Join(gopath, "pkg", "mod")
-	}
-
-	return gopath, gomodcache
-}
-
-// pinGoCacheForTestHome keeps module cache out of temp HOME directories in the test process.
-func pinGoCacheForTestHome(t *testing.T) {
-	t.Helper()
-	gopath, gomodcache := realGoCacheEnv(t)
-	t.Setenv("GOPATH", gopath)
-	t.Setenv("GOMODCACHE", gomodcache)
-}
-
-// cliChildEnv builds subprocess env with isolated HOME for credentials and pinned Go cache paths.
-func cliChildEnv(t *testing.T, testHome string) []string {
-	t.Helper()
-	gopath, gomodcache := realGoCacheEnv(t)
-
-	childEnv := make([]string, 0, len(os.Environ())+4)
-	for _, entry := range os.Environ() {
-		if strings.HasPrefix(entry, "HOME=") ||
-			strings.HasPrefix(entry, "USERPROFILE=") ||
-			strings.HasPrefix(entry, "GOPATH=") ||
-			strings.HasPrefix(entry, "GOMODCACHE=") {
-			continue
-		}
-		childEnv = append(childEnv, entry)
-	}
-	childEnv = append(childEnv,
-		"HOME="+testHome,
-		"USERPROFILE="+testHome,
-		"GOPATH="+gopath,
-		"GOMODCACHE="+gomodcache,
-	)
-	return childEnv
-}
-
 func createTestJWT(orgID string) string {
 	return testjwt.CreateTestJWT(orgID)
 }
@@ -297,7 +244,7 @@ func workflowDeployPrivateRegistry(t *testing.T, tc TestConfig) string {
 
 	cmd := exec.Command(CLIPath, args...)
 	testHome := CreateTestBearerCredentialsHome(t)
-	cmd.Env = cliChildEnv(t, testHome)
+	cmd.Env = testutil.CLIChildEnv(t, testHome)
 
 	var stdout, stderr bytes.Buffer
 	cmd.Stdout, cmd.Stderr = &stdout, &stderr
@@ -456,7 +403,7 @@ func workflowPausePrivateRegistry(t *testing.T, tc TestConfig) string {
 
 	cmd := exec.Command(CLIPath, args...)
 	testHome := CreateTestBearerCredentialsHome(t)
-	cmd.Env = cliChildEnv(t, testHome)
+	cmd.Env = testutil.CLIChildEnv(t, testHome)
 
 	var stdout, stderr bytes.Buffer
 	cmd.Stdout, cmd.Stderr = &stdout, &stderr
@@ -611,7 +558,7 @@ func workflowActivatePrivateRegistry(t *testing.T, tc TestConfig) string {
 
 	cmd := exec.Command(CLIPath, args...)
 	testHome := CreateTestBearerCredentialsHome(t)
-	cmd.Env = cliChildEnv(t, testHome)
+	cmd.Env = testutil.CLIChildEnv(t, testHome)
 
 	var stdout, stderr bytes.Buffer
 	cmd.Stdout, cmd.Stderr = &stdout, &stderr
@@ -754,7 +701,7 @@ func workflowDeletePrivateRegistry(t *testing.T, tc TestConfig) string {
 
 	cmd := exec.Command(CLIPath, args...)
 	testHome := CreateTestBearerCredentialsHome(t)
-	cmd.Env = cliChildEnv(t, testHome)
+	cmd.Env = testutil.CLIChildEnv(t, testHome)
 
 	var stdout, stderr bytes.Buffer
 	cmd.Stdout, cmd.Stderr = &stdout, &stderr
@@ -846,7 +793,7 @@ func RunPrivateRegistryAuthAndSettingsFinalize(t *testing.T, envPath, blankWorkf
 	bearerHome := CreateTestBearerCredentialsHome(t)
 	t.Setenv("HOME", bearerHome)
 	t.Setenv("USERPROFILE", bearerHome)
-	pinGoCacheForTestHome(t)
+	testutil.PinGoCacheForTestHome(t)
 
 	logger := testutil.NewTestLogger()
 	creds, err := credentials.New(logger)
